+++ NEWS.md
@@ -1,6 +1,6 @@
 ## ggimage 0.1.1
 
-+ annotating phylogenetic tree with ggtree and ggimage vignette <2017-12-21, Thu>
++ add angle parameter <2017-12-27, Wed>
 
 CHANGES IN VERSION 0.1.0
 ------------------------